protocols/backend/election-protocols-be/pyproject.toml
[project]
name = "election-protocols-be"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "uvicorn[standard]",
    "gunicorn",
    "fastapi",
    "pydantic-settings",
    "python-dotenv",
    "python-multipart",
    "google-genai"
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/election_protocols_be"]
